chore(refactor & revamp): impement > UI Layout > backend implementation
