+++ server/middlewares/authenticateUser.js
