+++ ui/admin/adminDashboard.html
@@ -11,13 +11,13 @@
     <div class="sidebar">
         <div class="sidebarItems">
             <div id="innerNav">
-                    <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </div>
                     <img src="../assets/images/profileImg.jpg" alt="Profile Image">
                     <div><a href="adminDashboard.html">Dashboard</a></div>
                     <div><a href="allBankAccounts.html">User accounts</a></div>
                             <div><a href="allBankTransactions.html"> Bank Transactions</a></div>
-                            <div><a href="createAdminAccount.html">Create admin account</a></div>
+                            <div><a href="createAdminAccount.html">Create admin</a></div>
                             <div><a href="changepassword.html">Change password</a></div>
                             <div><a href="../index.html">Logout</a></div>
     </div>
@@ -25,7 +25,7 @@
     <div class="userAccountCoverPage">
     <nav class="landingProfile">
         <div><h4>
-            banka.com</h4>
+            banka</h4>
         </div>
         <div>
                 <i class="fas fa-bars" id="menuLogo"></i>
@@ -58,7 +58,7 @@
                                         <td>Obayomi</td>
                                         <td>$270.00</td>
                                         <td>Mar/20/2019</td>
-                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Credit</td>
@@ -66,7 +66,7 @@
                                         <td>Obayomi</td>
                                         <td>$270.00</td>
                                         <td>Mar/20/2019</td>
-                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Credit</td>
@@ -74,7 +74,7 @@
                                         <td>Obayomi</td>
                                         <td>$270.00</td>
                                         <td>Mar/20/2019</td>
-                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Credit</td>
@@ -82,7 +82,7 @@
                                         <td>Obayomi</td>
                                         <td>$270.00</td>
                                         <td>Mar/20/2019</td>
-                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Credit</td>
@@ -90,7 +90,7 @@
                                         <td>Obayomi</td>
                                         <td>$270.00</td>
                                         <td>Mar/20/2019</td>
-                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     
                                 </table>
@@ -118,7 +118,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>May/12/2019</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>2</td>
@@ -127,7 +127,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>May/01/2019</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>3</td>
@@ -136,7 +136,7 @@
                                         <td>Savings</td>
                                         <td>Active</td>
                                         <td>Feb/20/2019</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>4</td>
@@ -145,7 +145,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>Jan/17/2019</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>5</td>
@@ -154,7 +154,7 @@
                                         <td>Savings</td>
                                         <td>Active</td>
                                         <td>Dec/06/2018</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr> 
                                 </table> 
                             </div>
@@ -162,7 +162,6 @@
         </div>
     </div>
 </div>
-   <!-- <footer class="footer">Andela Developer Challenge - &copy 2019</footer> -->
 <script src="../script/admin/adminAccount.js"></script>
 </body>
 </html>
