+++ ui/admin/allBankAccounts.html
@@ -11,13 +11,13 @@
     <div class="sidebar">
         <div class="sidebarItems">
             <div id="innerNav">
-                    <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </div>
                     <img src="../assets/images/profileImg.jpg" alt="Profile Image">
                     <div><a href="adminDashboard.html">Dashboard</a></div>
                     <div><a href="allBankAccounts.html">User accounts</a></div>
                     <div><a href="allBankTransactions.html">Bank Transactions</a></div>
-                            <div><a href="createAdminAccount.html">Create admin account</a></div>
+                            <div><a href="createAdminAccount.html">Create admin</a></div>
                             <div><a href="changepassword.html">Change password</a></div>
                             <div><a href="../index.html">Logout</a></div>
     </div>
@@ -25,7 +25,7 @@
     <div class="userAccountCoverPage">
     <nav class="landingProfile">
         <div><h4>
-            banka.com</h4>
+            banka</h4>
         </div>
         <div>
                 <i class="fas fa-bars" id="menuLogo"></i>
@@ -60,7 +60,7 @@
                                         <td>Savings</td>
                                         <td>Active</td>
                                         <td>May/12/2017</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>2</td>
@@ -69,7 +69,7 @@
                                         <td>Savings</td>
                                         <td>Inactive</td>
                                         <td>May/12/2017</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>3</td>
@@ -78,7 +78,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>Jan/02/2017</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>4</td>
@@ -87,7 +87,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>Jul/22/2016</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>5</td>
@@ -96,7 +96,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>Mar/24/2017</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>6</td>
@@ -105,7 +105,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>May/12/2017</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>7</td>
@@ -114,7 +114,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>May/12/2017</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>8</td>
@@ -123,7 +123,7 @@
                                         <td>Savings</td>
                                         <td>Inactive</td>
                                         <td>Sep/01/2016</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>9</td>
@@ -132,7 +132,7 @@
                                         <td>Current</td>
                                         <td>Active</td>
                                         <td>Sep/17/2018</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                             <td>10</td>
@@ -141,7 +141,7 @@
                                         <td>Savings</td>
                                         <td>Active</td>
                                         <td>Jun/06/2016</td>
-                                        <td><a href="bankAccountRecord.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="bankAccountRecord.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                 </table>       
                                 <div class="pagination">
@@ -156,7 +156,6 @@
     </div>
     
 </div>
-<!-- <footer class="footer">Andela Developer Challenge - &copy 2019</footer> -->
 <script src="../script/admin/adminAccount.js"></script>
 </body>
 </html>
