+++ ui/admin/bankAccountRecord.html
@@ -11,13 +11,13 @@
     <div class="sidebar">
         <div class="sidebarItems">
             <div id="innerNav">
-                    <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </div>
                     <img src="../assets/images/profileImg.jpg" alt="Profile Image">
                     <div><a href="adminDashboard.html">Dashboard</a></div>
                     <div><a href="allBankAccounts.html">User accounts</a></div>
                     <div><a href="allBankTransactions.html">Bank Transactions</a></div>
-                    <div><a href="createAdminAccount.html">Create admin account</a></div>
+                    <div><a href="createAdminAccount.html">Create admin</a></div>
                     <div><a href="changepassword.html">Change password</a></div>
                     <div><a href="../index.html">Logout</a></div>
     </div>
@@ -25,7 +25,7 @@
     <div class="userAccountCoverPage">
         <nav class="landingProfile">
             <div><h4>
-                banka.com</h4>
+                banka</h4>
             </div>
             <div>
                     <i class="fas fa-bars" id="menuLogo"></i>
@@ -38,13 +38,12 @@
                         </div>
                  <div class="ownerInfo">
                      <ul>
-                         <li>Account name: Obayomi Martins O.</li>
-                         <li>Account number: 22399029393</li>
-                         <li>Account type: Savings</li>
-                         <li>Account status: Active</li>
-                         <li>Current balance: $3,550.05</li>
-                         <li>Account email: obayomi@gmail.com</li>
-                         <li>Telephone: 07084929485</li>
+                         <li><b>Account name:</b> Obayomi Martins O.</li>
+                         <li><b>Account number:</b> 22399029393</li>
+                         <li><b>Account type:</b> Savings</li>
+                         <li><b>Account status:</b> Active</li>
+                         <li><b>Current balance:</b> $3,550.05</li>
+                         <li><b>Account email:</b> obayomi@gmail.com</li>
                      </ul>
                  </div>
              </div>
@@ -81,7 +80,7 @@
                                             <td>Obayomi</td>
                                             <td>$270.00</td>
                                             <td>Mar/20/2019</td>
-                                            <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                            <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></i></a></td>
                                         </tr>
                                         <tr>
                                             <td>Debit</td>
@@ -89,7 +88,7 @@
                                             <td>joan</td>
                                             <td>$185.00</td>
                                             <td>Feb/12/2019</td>
-                                            <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                            <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                         </tr>
                                         <tr>
                                                 <td>Credit</td>
@@ -97,7 +96,7 @@
                                                 <td>Obayomi</td>
                                                 <td>$25.50</td>
                                                 <td>Jan/29/2019</td>
-                                                <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                                <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                             </tr>
                                             <tr>
                                                     <td>Debit</td>
@@ -105,7 +104,7 @@
                                                     <td>Nelson</td>
                                                     <td>$75.00</td>
                                                     <td>Jan/02/2019</td>
-                                                    <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                                    <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                                 </tr>
                                         <tr>
                                             <td>Credit</td>
@@ -113,7 +112,7 @@
                                             <td>Obayomi</td>
                                             <td>$100.00 </td>
                                             <td>Dec/25/2018</td>
-                                            <td><a href="../user/transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                            <td><a href="../user/transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                         </tr> 
                                     </table>       
                                     <div class="pagination">
@@ -134,12 +133,6 @@
                                             </div>
                                         </div>
                                         <div class="deleteAccountModal deleteAccountModal2">
-                                            <!-- <div class="deleteModalContent deleteModalContent2">
-                                                <div class="confirmBtn">
-                                                    <button class="confirmbtn" id="deleteAccountBtn">Yes</button>
-                                                    <button class="confirmbtn" id="cancelDeleteBtn">Cancel</button>
-                                                </div>
-                                            </div> -->
                                         </div>
                         </div>
                     </div>
