+++ ui/admin/changepassword.html
@@ -11,7 +11,7 @@
     <nav id="navbar" class="landing">
         <div id="nav-container">                
             <h3 class="nav-title">
-                <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </h3>
         </div>
     </nav>
@@ -28,10 +28,10 @@
                 <div> 
                         <input type="password" name="password" class="formInput" placeholder="New password confirm">
                     </div>
-                <div class="loginBtn"><a href="index.html">Change password</a></div>
+                <div class="loginBtn"><a href="adminDashboard.html">Change password</a></div>
                 <div class="innerFooter">
                     <span class="reset"><a href="../resetPassword.html">RESET PASSWORD</a></span>
-                    <span class="create"><a href="userAccount.html">MY ACCOUNT</a></span>
+                    <span class="create"><a href="adminDashboard.html">ADMIN DASHBOARD</a></span>
                 </div>
             </form>
         </div>
