+++ ui/admin/createAdminAccount.html
@@ -4,20 +4,20 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Banka</title>
-    <link type="text/css" rel="stylesheet" href="assets/css/index.css">
+    <link type="text/css" rel="stylesheet" href="../assets/css/index.css">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.7.2/css/all.css" integrity="sha384-fnmOCqbTlWIlj8LyTjo7mOUStjsKC4pOpQbqyi7RrhN7udi9RwhKkMHpvLbHG9Sr" crossorigin="anonymous">
 </head>
 <body>
     <nav id="navbar" class="landing">
         <div id="nav-container">                
             <h3 class="nav-title">
-                <i class="fas fa-piggy-bank" id="logo"></i>
+                <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </h3>
         </div>
     </nav>
     <div class="loginCoverPage">
         <div class="formModal">
-            <a href="index.html"><h1>banka.com</h1></a>
+            <a href="#"><h1>Create admin account</h1></a>
             <form action="" id="loginForm">
                 <div>
                     <input type="text" value="" class="formInput" placeholder="Firstname">   
@@ -29,33 +29,26 @@
                         <input type="email" value="" class="formInput" placeholder="Email">   
                     </div>
                     <div>
-                            <input type="number" value="" class="formInput" placeholder="Your Age">
-                        </div>
-                    <div>
-                        <select name="Account type" id="">
-                            <option value="Savings">Gender</option>
-                            <option value="Savings">Male</option>
-                            <option value="Current">Female</option>
-                        </select>
+                        <select name="Admin type">
+                                <option value="Admin">Admin or Staff account</option>
+                                <option value="Admin">Admin account</option>
+                                <option value="Staff">Staff account</option>
+                            </select>
                     </div>
-                                            <div id="img">
-                            <label for="Image">Image upload<sup>*</sup></label>
-                                <input type="file" value="Image" class="formInput" accept="image/*">
-                            </div>
                 <div> 
-                    <input type="password" name="password" class="formInput" placeholder="Enter password">
+                    <input type="password" name="password" class="formInput" placeholder="Admin password">
                 </div>
                 <div> 
                         <input type="password" name="password" class="formInput" placeholder="Password Confirm">
                     </div>
-                <div class="loginBtn"><a href="index.html">Sign Up</a></div>
+                <div class="loginBtn"><a href="index.html">Create account</a></div>
                 <div class="innerFooter">
-                    <span class="reset"><a href="resetPassword.html">RESET PASSWORD</a></span>
-                    <span class="create"><a href="index.html">SIGN IN</a></span>
+                    <span class="reset"><a href="../index.html">LOGOUT</a></span>
+                    <span class="create"><a href="adminDashboard.html">ADMIN DASHBOARD</a></span>
                 </div>
             </form>
         </div>
-        <footer>Andela Developer Challenge - @ 2019</footer>
     </div>
+    <footer>Andela Developer Challenge - &copy 2019</footer>
 </body>
 </html>
