+++ ui/index.html
@@ -12,13 +12,13 @@
     <nav id="navbar" class="landing">
         <div id="nav-container">                
             <h3 class="nav-title">
-                <i class="fas fa-piggy-bank" id="logo"></i>
+                <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </h3>
         </div>
     </nav>
     <div class="loginCoverPage">
         <div class="formModal">
-            <a href="index.html"><h1>banka.com</h1></a>
+            <a href="index.html"><h1>banka</h1></a>
             <form action="" id="loginForm">
                 <div>
                     <input type="text" value="" class="formInput" placeholder="Enter your email" required>   
@@ -30,11 +30,12 @@
                 <div class="innerFooter">
                     <span class="reset"><a href="resetPassword.html">RESET PASSWORD</a></span>
                     <span class="create"><a href="signup.html">CREATE AN ACCOUNT</a></span>
+                    <div class="adminLogin"><a href="adminLogin.html">ADMIN LOGIN</a></div>
                 </div>
             </form>
         </div>
         <hr>
     </div>
-    <footer>Andela Developer Challenge - @ 2019</footer>
+    <footer>Andela Developer Challenge - &copy 2019</footer>
 </body>
 </html>
