+++ ui/resetPassword.html
@@ -11,7 +11,7 @@
     <nav id="navbar" class="landing">
         <div id="nav-container">                
             <h3 class="nav-title">
-                <i class="fas fa-piggy-bank" id="logo"></i>
+                <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </h3>
         </div>
     </nav>
@@ -30,6 +30,6 @@
             </form>
         </div>
     </div>
-    <footer>Andela Developer Challenge - @ 2019</footer>
+    <footer>Andela Developer Challenge - &copy 2019</footer>
 </body>
 </html>
