+++ ui/user/createBankAccount.html
@@ -11,17 +11,14 @@
     <nav id="navbar" class="landing">
         <div id="nav-container">                
             <h3 class="nav-title">
-                <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </h3>
         </div>
     </nav>
     <div class="createBankAccountCoverPage">
         <div class="formModal">
-            <h4>Create a banka.com account</h4>
+            <h4>Create a banka account</h4>
             <form action="" id="loginForm">
-                <div>
-                        <input type="text" value="" class="formInput" placeholder="Account Number" required>   
-                    </div>
                 <div>
                         <select name="Account type" id="">
                             <option value="type">Account type</option>
@@ -29,9 +26,6 @@
                             <option value="Current">Savings account</option>
                         </select>
                     </div>
-                    <div>
-                        <input type="email" value="" class="formInput" placeholder="Email" required>   
-                        </div>
                     <div>
                             <input type="number" value="" class="formInput" placeholder="Initial deposit" required>   
                         </div>
