+++ ui/user/transactionDetails.html
@@ -11,10 +11,9 @@
     <div class="sidebar">
         <div class="sidebarItems">
             <div id="innerNav">
-                    <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </div>
                     <img src="../assets/images/profileImg.jpg" alt="Profile Image">
-                    <!-- <small>ID: OB5466</small> -->
                    <div><a href="userAccounttransactions.html">My Transactions</a></div>
                             <div><a href="userAccount.html">My account</a></div>
                             <div><a href="changePassword.html">Change password</a></div>
@@ -25,7 +24,7 @@
     <div class="transactionPage">
     <nav class="landingProfile">
         <div><h4>
-            banka.com</h4>
+            banka</h4>
         </div>
         <div>
                 <i class="fas fa-bars" id="menuLogo"></i>
@@ -69,7 +68,6 @@
             </div>
     </div>
     </div>
-     <!-- <footer class="footer">Andela Developer Challenge - &copy 2019</footer> -->
 <script src="../script/admin/adminAccount.js"></script>
 </body>
 </html>
