+++ ui/user/userAccount.html
@@ -11,10 +11,9 @@
     <div class="sidebar">
         <div class="sidebarItems">
             <div id="innerNav">
-                    <i class="fas fa-piggy-bank" id="logo"></i>
+                    <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </div>
                     <img src="../assets/images/profileImg.jpg" alt="Profile Image">
-                    <!-- <small>ID: OB5466</small> -->
                             <div><a href="userAccounttransactions.html">My Transactions</a></div>
                             <div><a href="userAccount.html">My account</a></div>
                             <div><a href="changePassword.html">Change password</a></div>
@@ -25,16 +24,12 @@
     <div class="userAccountCoverPage">
     <nav class="landingProfile">
         <div><h4>
-            banka.com</h4>
+            banka</h4>
         </div>
         <div>
                 <i class="fas fa-bars" id="menuLogo"></i>
         </div>   
     </nav>
-
-    <!-- <div class="userInfo">
-        <h4>USER DETAILS</h4>
-    </div> -->
      <div id="accountSection">
          <div class="details">
         <ul>
@@ -81,7 +76,7 @@
                                         <td>Obayomi</td>
                                         <td>$270.00</td>
                                         <td>Mar/20/2019</td>
-                                        <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Debit</td>
@@ -89,7 +84,7 @@
                                         <td>joan</td>
                                         <td>$185.00</td>
                                         <td>Mar/02/2019</td>
-                                        <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Debit</td>
@@ -97,14 +92,13 @@
                                         <td>banka.com</td>
                                         <td>$14.07 </td>
                                         <td>Dec/26/2018</td>
-                                        <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr> 
                                 </table>       
                     </div>
                 </div>
         </div>
     </div>
-    <!-- <footer class="footer">Andela Developer Challenge - &copy 2019</footer> -->
 </div>
 <script src="../script/user/userAccount.js"></script>
 </body>
