+++ ui/user/userAccounttransactions.html
@@ -11,10 +11,9 @@
     <div class="sidebar">
         <div class="sidebarItems">
             <div id="innerNav">
-                    <i class="fas fa-piggy-bank" id="logo"></i>
+                <a href="index.html"><i class="fas fa-piggy-bank" id="logo"></i></a>
             </div>
                     <img src="../assets/images/profileImg.jpg" alt="Profile Image">
-                    <!-- <small>ID: OB5466</small> -->
                             <div><a href="userAccounttransactions.html">My Transactions</a></div>
                             <div><a href="userAccount.html">My account</a></div>
                             <div><a href="changePassword.html">Change password</a></div>
@@ -25,7 +24,7 @@
     <div class="transactionPage">
     <nav class="landingProfile">
         <div><h4>
-            banka.com</h4>
+            banka</h4>
         </div>
         <div>
                 <i class="fas fa-bars" id="menuLogo"></i>
@@ -42,12 +41,12 @@
                 Filter: <input type="date" name="from"> to: <input type="date" name="to"></div>
         </div>
         <div id="accountSection">
+                <table class="accBal">
+                        <th>Current balance</th>
+                        <th>$3237.06</th>
+                    </tr>
+                </table> 
                 <div class="bankAccountInfo user">
-                        <table class="accBal">
-                                    <th>Current balance</th>
-                                    <th>$3237.06</th>
-                                </tr>
-                            </table> 
                         <table>
                                 
                                 <tr>
@@ -64,7 +63,7 @@
                                     <td>Obayomi</td>
                                     <td>$270.00</td>
                                     <td>Mar/20/2019</td>
-                                    <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                    <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                 </tr>
                                 <tr>
                                     <td>Debit</td>
@@ -72,7 +71,7 @@
                                     <td>joan</td>
                                     <td>$185.00</td>
                                     <td>Mar/02/2019</td>
-                                    <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                    <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                 </tr>
                                 <tr>
                                     <td>Debit</td>
@@ -80,7 +79,7 @@
                                     <td>banka.com</td>
                                     <td>$14.07 </td>
                                     <td>Dec/26/2018</td>
-                                    <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                    <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                 </tr>
                                 <tr>
                                         <td>Debit</td>
@@ -88,7 +87,7 @@
                                         <td>Obayomi</td>
                                         <td>$67.99 </td>
                                         <td>Dec/04/2018</td>
-                                        <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Credit</td>
@@ -96,7 +95,7 @@
                                         <td>Obayomi</td>
                                         <td>$700.00 </td>
                                         <td>Oct/24/2018</td>
-                                        <td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+                                        <td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                     </tr>
                                     <tr>
                                         <td>Credit</td>
@@ -104,7 +103,7 @@
                                         <td>Obayomi</td>
                                         <td>$2000.00 </td>
                                         <td>April/11/2018</td>
-<td><a href="transactionDetails.html"><i class="fas fa-info-circle" id="tableInfo"></i></a></td>
+<td><a href="transactionDetails.html"><i class="fas fa-eye" id="tableInfo"></i></a></td>
                                         </tr> 
                             </table>  
                             <div class="pagination">
@@ -116,7 +115,6 @@
                
             </div>
     </div>
-      <!-- <footer class="footer">Andela Developer Challenge - &copy 2019</footer> -->
 <script src="../script/user/userAccount.js"></script>
 </body>
 </html>
